fix(ci): 修复旧 CI sanitizer 和 warning
PR 分支此前仍有 sanitizer 启动阶段失败和 clang warning backlog，本次按旧 CI 日志逐项收口，不改 Windows 构建入口。

## fix

- 防御资源页卡片 cache key 的非有限 UiScale/CardWidth，避免 ASan/UBSan 在 round_to_int(-inf) 上失败。

- 清理旧 clang warning 清单中的 unused、shadow、sign-compare、format-nonliteral、thread-safety 诊断。

- 保留监控测试需要的 stable-text sentinel，并同步源码审计断言。

## docs

- 将 Rust 测试验证命令统一为 run_rust_tests -j 14；default gate 已按该口径执行。

## test

- qmclient_scripts/cmake-windows.cmd --build cmake-build-release --target game-client -j 14

- qmclient_scripts/cmake-windows.cmd --build cmake-build-release --target run_cxx_tests -j 14（1439/1439）

- qmclient_scripts/cmake-windows.cmd --build cmake-build-release --target run_rust_tests -j 14

- python qmclient_scripts/languages_qmclient/extract_strings.py && python qmclient_scripts/languages_qmclient/generate_all.py && python qmclient_scripts/languages_qmclient/validate.py && python qmclient_scripts/languages_qmclient/review_duplicate_entries.py --show-groups 0 --show-unused 0

- python qmclient_scripts/gate/check_docs.py --sync-only --prefer agents && python qmclient_scripts/gate/check_docs.py

- python qmclient_scripts/gate/check_gate.py --mode quick --report-json-path tmp/check-gate-report.json

- python qmclient_scripts/gate/check_gate.py --mode default --report-json-path tmp/check-gate-report-default.json

+++ src/game/client/components/chat.h
@@ -351,7 +351,7 @@ public:
 
 		const size_t WordStart = str_length(pList);
 		str_append(pList, pWord, ListSize);
-		return str_length(pList) > WordStart;
+		return (size_t)str_length(pList) > WordStart;
 	}
 	static bool BuildWhisperCommand(char *pBuf, size_t BufSize, const char *pName, const char *pMessage)
 	{
